Repair job acquires Consul lock before repair
It is anticipated that all nodes of a cluster attempting to repair a
particular keyspace will compete for the same lock around the same time.

The repair program will block until it acquires the lock. Only 1 lock
can be held at a time. Each node will therefore be repaired one after
the other.

This iteration does not yet anticipate potential session failure. In
which case the lock would be released prematurely.