Terminate repair process if Consul lock is lost
This is highly unlikely to happen, but hypothetically possible. If a
session were to expire for example, the lock would be lost. This could
happen if there's a problem with the local Consul node and it was not
able to be renewed before its expiration (sessions have a 30s TTL).

In the case that this happens, we'd like to terminate the repair process
running. This is so if another node acquires the lock, we do not have
multiple nodes being repaired simultaneously.