Run dc-parallel repair instead of sequential
The initial design anticipated sequential repair of all nodes in all
datacenters. It was realized in the implementation process that Consul
does not support locking among multiple datacenters. Nodes can only
create sessions/locks in the datacenter they are registered in.

Instead there will be a coordinating datacenter that will be repaired
sequentially. All other datacenters will be repaired in parallel in the
sequence of the coordinating datacenter.